[ISSUE #8166] optimize: make compression type configurable in producer clinet level
